requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.25
botocore==1.31.25
dj-database-url==2.0.0
Django==4.2
django-compat==1.0.15
django-cors-headers==4.2.0
django-db-connection-pool==1.2.4
django-hijack==3.0.0
django-multiselectfield==0.1.12
django-storages==1.13.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
greenlet==3.0.1
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.22
sqlparams==5.1.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
django-filter==24.2
